refactor(structured): extract structured output helpers
