+++ src/JKToolKit.CodexSDK/StructuredOutputs/CodexStructuredOutputExtensions.cs
@@ -1,11 +1,9 @@
-using System.Text;
 using System.Text.Json;
 using JKToolKit.CodexSDK.Abstractions;
 using JKToolKit.CodexSDK.AppServer;
-using JKToolKit.CodexSDK.AppServer.Notifications;
 using JKToolKit.CodexSDK.Exec;
-using JKToolKit.CodexSDK.Exec.Notifications;
 using JKToolKit.CodexSDK.Exec.Protocol;
+using JKToolKit.CodexSDK.StructuredOutputs.Internal;
 
 namespace JKToolKit.CodexSDK.StructuredOutputs;
 
@@ -36,8 +34,8 @@ public static class CodexStructuredOutputExtensions
 
         await using var session = await client.StartSessionAsync(effective, ct).ConfigureAwait(false);
 
-        var raw = await CaptureExecFinalTextAsync(session, EventStreamOptions.Default, ct).ConfigureAwait(false);
-        var (value, json) = DeserializeStructured<T>(raw, structured, serializerOptions);
+        var raw = await StructuredOutputExecCapture.CaptureExecFinalTextAsync(session, EventStreamOptions.Default, ct).ConfigureAwait(false);
+        var (value, json) = StructuredOutputDeserializer.DeserializeStructured<T>(raw, structured, serializerOptions);
 
         return new CodexStructuredResult<T>
         {
@@ -80,8 +78,8 @@ public static class CodexStructuredOutputExtensions
 
         await using var session = await client.ResumeSessionAsync(sessionId, effective, ct).ConfigureAwait(false);
 
-        var raw = await CaptureExecFinalTextAsync(session, EventStreamOptions.FromTimestamp(resumeStart, follow: true), ct).ConfigureAwait(false);
-        var (value, json) = DeserializeStructured<T>(raw, structured, serializerOptions);
+        var raw = await StructuredOutputExecCapture.CaptureExecFinalTextAsync(session, EventStreamOptions.FromTimestamp(resumeStart, follow: true), ct).ConfigureAwait(false);
+        var (value, json) = StructuredOutputDeserializer.DeserializeStructured<T>(raw, structured, serializerOptions);
 
         return new CodexStructuredResult<T>
         {
@@ -121,8 +119,8 @@ public static class CodexStructuredOutputExtensions
 
         await using var turn = await client.StartTurnAsync(threadId, effective, ct).ConfigureAwait(false);
 
-        var raw = await CaptureAppServerFinalTextAsync(turn, ct).ConfigureAwait(false);
-        var (value, json) = DeserializeStructured<T>(raw, structured, serializerOptions);
+        var raw = await StructuredOutputAppServerCapture.CaptureAppServerFinalTextAsync(turn, ct).ConfigureAwait(false);
+        var (value, json) = StructuredOutputDeserializer.DeserializeStructured<T>(raw, structured, serializerOptions);
 
         return new CodexStructuredResult<T>
         {
@@ -150,9 +148,9 @@ public static class CodexStructuredOutputExtensions
         ArgumentNullException.ThrowIfNull(options);
         ArgumentNullException.ThrowIfNull(progress);
 
-        var prepared = PrepareExecStructuredRetry<T>(options, retry, structured);
+        var prepared = StructuredOutputRetryRunner.PrepareExecStructuredRetry<T>(options, retry, structured);
 
-        return await RunStructuredWithRetryCoreAsync<T>(
+        return await StructuredOutputRetryRunner.RunExecStructuredWithRetryCoreAsync<T>(
             client,
             initialSessionId: null,
             prepared.EffectiveBase,
@@ -176,9 +174,9 @@ public static class CodexStructuredOutputExtensions
         ArgumentNullException.ThrowIfNull(client);
         ArgumentNullException.ThrowIfNull(options);
 
-        var prepared = PrepareExecStructuredRetry<T>(options, retry, structured);
+        var prepared = StructuredOutputRetryRunner.PrepareExecStructuredRetry<T>(options, retry, structured);
 
-        return await RunStructuredWithRetryCoreAsync<T>(
+        return await StructuredOutputRetryRunner.RunExecStructuredWithRetryCoreAsync<T>(
             client,
             initialSessionId: null,
             prepared.EffectiveBase,
@@ -210,9 +208,9 @@ public static class CodexStructuredOutputExtensions
             throw new ArgumentException("SessionId cannot be empty or whitespace.", nameof(sessionId));
         }
 
-        var prepared = PrepareExecStructuredRetry<T>(options, retry, structured);
+        var prepared = StructuredOutputRetryRunner.PrepareExecStructuredRetry<T>(options, retry, structured);
 
-        return await RunStructuredWithRetryCoreAsync<T>(
+        return await StructuredOutputRetryRunner.RunExecStructuredWithRetryCoreAsync<T>(
             client,
             sessionId,
             prepared.EffectiveBase,
@@ -241,9 +239,9 @@ public static class CodexStructuredOutputExtensions
             throw new ArgumentException("SessionId cannot be empty or whitespace.", nameof(sessionId));
         }
 
-        var prepared = PrepareExecStructuredRetry<T>(options, retry, structured);
+        var prepared = StructuredOutputRetryRunner.PrepareExecStructuredRetry<T>(options, retry, structured);
 
-        return await RunStructuredWithRetryCoreAsync<T>(
+        return await StructuredOutputRetryRunner.RunExecStructuredWithRetryCoreAsync<T>(
             client,
             sessionId,
             prepared.EffectiveBase,
@@ -252,143 +250,6 @@ public static class CodexStructuredOutputExtensions
             prepared.SerializerOptions,
             progress: null,
             ct).ConfigureAwait(false);
-    }
-
-    private static (
-        CodexSessionOptions EffectiveBase,
-        CodexStructuredRetryOptions Retry,
-        CodexStructuredOutputOptions Structured,
-        JsonSerializerOptions SerializerOptions) PrepareExecStructuredRetry<T>(
-        CodexSessionOptions options,
-        CodexStructuredRetryOptions? retry,
-        CodexStructuredOutputOptions? structured)
-    {
-        retry ??= new CodexStructuredRetryOptions();
-        structured ??= new CodexStructuredOutputOptions();
-
-        if (retry.MaxAttempts <= 0)
-        {
-            throw new ArgumentOutOfRangeException(nameof(retry.MaxAttempts), retry.MaxAttempts, "MaxAttempts must be greater than zero.");
-        }
-
-        var serializerOptions = structured.SerializerOptions ?? new JsonSerializerOptions(JsonSerializerDefaults.Web);
-        var schema = CodexJsonSchemaGenerator.Generate<T>(serializerOptions);
-
-        var effectiveBase = options.Clone();
-        effectiveBase.OutputSchema = CodexOutputSchema.FromJson(schema);
-
-        return (effectiveBase, retry, structured, serializerOptions);
-    }
-
-    private static async Task<CodexStructuredResult<T>> RunStructuredWithRetryCoreAsync<T>(
-        ICodexClient client,
-        SessionId? initialSessionId,
-        CodexSessionOptions effectiveBase,
-        CodexStructuredRetryOptions retry,
-        CodexStructuredOutputOptions structured,
-        JsonSerializerOptions serializerOptions,
-        CodexStructuredRunProgress? progress,
-        CancellationToken ct)
-    {
-        SessionId? sessionId = initialSessionId;
-        string? logPath = null;
-        CodexStructuredOutputParseException? lastParse = null;
-
-        for (var attempt = 1; attempt <= retry.MaxAttempts; attempt++)
-        {
-            ct.ThrowIfCancellationRequested();
-
-            var isStartAttempt = attempt == 1 && sessionId is null;
-            var attemptKind = isStartAttempt ? CodexStructuredAttemptKind.Start : CodexStructuredAttemptKind.Resume;
-            progress?.AttemptStarting?.Invoke(attempt, retry.MaxAttempts, attemptKind);
-
-            try
-            {
-                if (isStartAttempt)
-                {
-                    await using var session = await client.StartSessionAsync(effectiveBase, ct).ConfigureAwait(false);
-                    sessionId = session.Info.Id;
-                    logPath = session.Info.LogPath;
-                    progress?.SessionLocated?.Invoke(session.Info.Id, session.Info.LogPath);
-
-                    var raw = await CaptureExecFinalTextAsync(session, EventStreamOptions.Default, progress?.EventReceived, ct).ConfigureAwait(false);
-                    var (value, json) = DeserializeStructured<T>(raw, structured, serializerOptions);
-
-                    return new CodexStructuredResult<T>
-                    {
-                        Value = value,
-                        RawJson = json,
-                        RawText = raw,
-                        SessionId = session.Info.Id.Value,
-                        LogPath = session.Info.LogPath
-                    };
-                }
-
-                if (sessionId is null)
-                {
-                    throw new InvalidOperationException("Retry attempted without a captured session id.");
-                }
-
-                var resumeStart = DateTimeOffset.UtcNow;
-                var resumeOptions = effectiveBase;
-                if (!(attempt == 1 && initialSessionId is not null))
-                {
-                    var retryPrompt = retry.BuildRetryPrompt(new CodexStructuredRetryContext
-                    {
-                        Attempt = attempt - 1,
-                        MaxAttempts = retry.MaxAttempts,
-                        RawText = lastParse?.RawText ?? string.Empty,
-                        ExtractedJson = lastParse?.ExtractedJson,
-                        Exception = (Exception?)lastParse ?? new InvalidOperationException("Unknown parse failure."),
-                        SessionId = sessionId.Value,
-                        LogPath = logPath
-                    });
-
-                    resumeOptions = effectiveBase.Clone();
-                    resumeOptions.Prompt = retryPrompt;
-                }
-
-                await using var resumed = await client.ResumeSessionAsync(sessionId.Value, resumeOptions, ct).ConfigureAwait(false);
-                logPath = resumed.Info.LogPath;
-                progress?.SessionLocated?.Invoke(resumed.Info.Id, resumed.Info.LogPath);
-
-                var raw2 = await CaptureExecFinalTextAsync(resumed, EventStreamOptions.FromTimestamp(resumeStart, follow: true), progress?.EventReceived, ct).ConfigureAwait(false);
-                var (value2, json2) = DeserializeStructured<T>(raw2, structured, serializerOptions);
-
-                return new CodexStructuredResult<T>
-                {
-                    Value = value2,
-                    RawJson = json2,
-                    RawText = raw2,
-                    SessionId = resumed.Info.Id.Value,
-                    LogPath = resumed.Info.LogPath
-                };
-            }
-            catch (CodexStructuredOutputParseException ex)
-            {
-                lastParse = ex;
-                progress?.ParseFailed?.Invoke(attempt, ex);
-
-                if (attempt >= retry.MaxAttempts)
-                {
-                    throw new CodexStructuredOutputRetryFailedException(
-                        attempts: attempt,
-                        message: $"Failed to parse structured output after {attempt} attempts.",
-                        innerException: ex,
-                        sessionId: sessionId?.Value,
-                        logPath);
-                }
-
-                // Next loop iteration will resume with retry prompt.
-            }
-        }
-
-        throw new CodexStructuredOutputRetryFailedException(
-            attempts: retry.MaxAttempts,
-            message: $"Failed to parse structured output after {retry.MaxAttempts} attempts.",
-            innerException: (Exception?)lastParse ?? new InvalidOperationException("Unknown failure."),
-            sessionId: sessionId?.Value,
-            logPath);
     }
 
     /// <summary>
@@ -434,8 +295,8 @@ public static class CodexStructuredOutputExtensions
 
                     await using var turn = await client.StartTurnAsync(threadId, effective, ct).ConfigureAwait(false);
 
-                    var raw = await CaptureAppServerFinalTextAsync(turn, ct).ConfigureAwait(false);
-                    var (value, json) = DeserializeStructured<T>(raw, structured, serializerOptions);
+                    var raw = await StructuredOutputAppServerCapture.CaptureAppServerFinalTextAsync(turn, ct).ConfigureAwait(false);
+                    var (value, json) = StructuredOutputDeserializer.DeserializeStructured<T>(raw, structured, serializerOptions);
 
                     return new CodexStructuredResult<T>
                     {
@@ -463,8 +324,8 @@ public static class CodexStructuredOutputExtensions
 
                 await using var retryTurn = await client.StartTurnAsync(threadId, next, ct).ConfigureAwait(false);
 
-                var rawRetry = await CaptureAppServerFinalTextAsync(retryTurn, ct).ConfigureAwait(false);
-                var (retryValue, retryJson) = DeserializeStructured<T>(rawRetry, structured, serializerOptions);
+                var rawRetry = await StructuredOutputAppServerCapture.CaptureAppServerFinalTextAsync(retryTurn, ct).ConfigureAwait(false);
+                var (retryValue, retryJson) = StructuredOutputDeserializer.DeserializeStructured<T>(rawRetry, structured, serializerOptions);
 
                 return new CodexStructuredResult<T>
                 {
@@ -488,177 +349,4 @@ public static class CodexStructuredOutputExtensions
         throw (Exception?)lastParse ?? new InvalidOperationException("Unknown structured output retry failure.");
     }
 
-    private static (T Value, string RawJson) DeserializeStructured<T>(string rawText, CodexStructuredOutputOptions structured, JsonSerializerOptions serializerOptions)
-    {
-        var extractedJson = (string?)null;
-        try
-        {
-            extractedJson = CodexStructuredJsonExtractor.ExtractJson(rawText, structured.TolerantJsonExtraction);
-        }
-        catch (Exception ex)
-        {
-            throw new CodexStructuredOutputParseException(
-                message: $"Failed to extract JSON for structured output type '{typeof(T).FullName}'.",
-                rawText: rawText,
-                extractedJson: null,
-                innerException: ex);
-        }
-
-        try
-        {
-            var value = JsonSerializer.Deserialize<T>(extractedJson, serializerOptions);
-            if (value is null)
-            {
-                throw new InvalidOperationException("Deserialized value was null.");
-            }
-
-            return (value, extractedJson);
-        }
-        catch (Exception ex)
-        {
-            throw new CodexStructuredOutputParseException(
-                message: $"Failed to deserialize structured output into '{typeof(T).FullName}'.",
-                rawText: rawText,
-                extractedJson: extractedJson,
-                innerException: ex);
-        }
-    }
-
-    private static async Task<string> CaptureExecFinalTextAsync(
-        ICodexSessionHandle session,
-        EventStreamOptions streamOptions,
-        CancellationToken ct)
-    {
-        return await CaptureExecFinalTextAsync(session, streamOptions, onEvent: null, ct).ConfigureAwait(false);
-    }
-
-    private static async Task<string> CaptureExecFinalTextAsync(
-        ICodexSessionHandle session,
-        EventStreamOptions streamOptions,
-        Action<CodexEvent>? onEvent,
-        CancellationToken ct)
-    {
-        // For live sessions, rely on process exit rather than a specific event name like `task_complete`.
-        // Codex CLI output formats can evolve; the most robust completion signal is the process exiting.
-        if (session.IsLive && streamOptions.Follow)
-        {
-            using var progressCts = CancellationTokenSource.CreateLinkedTokenSource(ct);
-
-            var progressTask = Task.Run(async () =>
-            {
-                try
-                {
-                    await foreach (var evt in session.GetEventsAsync(streamOptions, progressCts.Token).ConfigureAwait(false))
-                    {
-                        onEvent?.Invoke(evt);
-                    }
-                }
-                catch (OperationCanceledException)
-                {
-                    // expected during shutdown
-                }
-            }, CancellationToken.None);
-
-            try
-            {
-                await session.WaitForExitAsync(ct).ConfigureAwait(false);
-            }
-            finally
-            {
-                progressCts.Cancel();
-                try { await progressTask.ConfigureAwait(false); } catch { /* best-effort */ }
-            }
-
-            // After exit, re-scan the log without following to deterministically capture the final message.
-            return await ReadFinalExecTextFromStreamAsync(session, streamOptions with { Follow = false }, onEvent: null, ct).ConfigureAwait(false);
-        }
-
-        // Non-live sessions (or callers who disabled follow) can just read to EOF.
-        var effective = streamOptions with { Follow = false };
-        return await ReadFinalExecTextFromStreamAsync(session, effective, onEvent, ct).ConfigureAwait(false);
-    }
-
-    private static async Task<string> ReadFinalExecTextFromStreamAsync(
-        ICodexSessionHandle session,
-        EventStreamOptions streamOptions,
-        Action<CodexEvent>? onEvent,
-        CancellationToken ct)
-    {
-        string? raw = null;
-        var isDone = false;
-        await foreach (var evt in session.GetEventsAsync(streamOptions, ct).ConfigureAwait(false))
-        {
-            onEvent?.Invoke(evt);
-            switch (evt)
-            {
-                case AgentMessageEvent msg:
-                    raw = msg.Text;
-                    break;
-                case TaskCompleteEvent complete:
-                    raw = complete.LastAgentMessage ?? raw;
-                    isDone = true;
-                    break;
-            }
-
-            if (isDone)
-            {
-                break;
-            }
-        }
-        raw ??= string.Empty;
-        if (string.IsNullOrWhiteSpace(raw))
-        {
-            throw new CodexStructuredOutputParseException(
-                message: "Codex did not emit a final message to parse as structured output.",
-                rawText: raw,
-                extractedJson: null,
-                innerException: new InvalidOperationException("No final message."));
-        }
-
-        return raw;
-    }
-
-    private static async Task<string> CaptureAppServerFinalTextAsync(CodexTurnHandle turn, CancellationToken ct)
-    {
-        var deltas = new StringBuilder();
-        string? fullText = null;
-        var done = false;
-
-        await foreach (var evt in turn.Events(ct).ConfigureAwait(false))
-        {
-            switch (evt)
-            {
-                case AgentMessageDeltaNotification d:
-                    deltas.Append(d.Delta);
-                    break;
-                case ItemCompletedNotification ic when string.Equals(ic.ItemType, "agentMessage", StringComparison.Ordinal):
-                    if (ic.Item.ValueKind == JsonValueKind.Object &&
-                        ic.Item.TryGetProperty("text", out var t) &&
-                        t.ValueKind == JsonValueKind.String)
-                    {
-                        fullText = t.GetString();
-                    }
-                    break;
-                case TurnCompletedNotification:
-                    done = true;
-                    break;
-            }
-
-            if (done)
-            {
-                break;
-            }
-        }
-        var raw = fullText ?? deltas.ToString();
-        if (string.IsNullOrWhiteSpace(raw))
-        {
-            throw new CodexStructuredOutputParseException(
-                message: "Codex did not emit a final agent message to parse as structured output.",
-                rawText: raw,
-                extractedJson: null,
-                innerException: new InvalidOperationException("No final message."));
-        }
-
-        return raw;
-    }
 }
